Revisando configurações de build
