+++ tsconfig.app.json
@@ -1,6 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "exclude": [
+    "sandbox.ts",
     "**/*.spec.ts"
   ]
 }
